% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age}
\alias{age}
\title{Age data}
\format{data frame
  \describe{
  \item{year}{year sampled}
  \item{location}{location and type of sample}
  \item{stock}{Stock group of sample}
  \item{x3-x78}{Number of fish sampled by total age}
  \item{n}{sample size}
  }}
\source{
various
}
\usage{
age
}
\description{
Multinomial age data for Susitna River Chinook salmon
}
\keyword{datasets}
