% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{as}
\alias{as}
\title{Aerial Survey data}
\format{list
 \describe{
  \item{Deshka}{A matrix of aerial survey data from the Deshka River}
  \item{East Susitna}{A matrix of aerial survey data from the East Susitna stock group}
  \item{Talkeetna}{A matrix of aerial survey data from the Talkeetna drainage}
  \item{Yentna}{A matrix of aerial survey data from the Yentna drainage}
  \item{Other}{A matrix of aerial survey data from Other areas}
  }}
\usage{
as
}
\description{
Aerial survey data by stock and tributary.
}
\keyword{datasets}
