% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chinBEGs}
\alias{chinBEGs}
\title{Statewide Chinook Salmon Smsy and goal ranges}
\format{data.frame
  \describe{
  \item{Region}{Sport Fish Region}
  \item{Stock}{Name of the Stock}
  \item{Smsy}{Smsy of the Stock}
  \item{lb}{Lower bound of the Goal}
  \item{ub}{Upper bound of the Goal}
  }}
\source{
Initially compiled by Steve F.
}
\usage{
data(dat_chinBEGs)
}
\description{
Statewide Chinook Salmon Smsy and goal ranges
}
\keyword{datasets}
