% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_countprofile}
\alias{get_countprofile}
\title{Creates a dataset for plotting OYP, ORP and EY plots based on aerial count data}
\usage{
get_countprofile(post_dat, stock_name, trib_name)
}
\arguments{
\item{post_dat}{An mcmc object with nodes lnalpha, beta, S.msy, lnalpha.c, p.S#, theta, and sigma.air from which you want 
to simulate SR relationships.}

\item{stock_name}{A character element specifying the stock of interest.}

\item{trib_name}{A character element specifying the tributary aerial data.}
}
\value{
A data.frame
}
\description{
This function creates a dataframe that can be used by plot_countprofile()
}
\examples{
get_ids()
profile <- get_countprofile(post, "East Susitna", "Willow")

}
