% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_Hhat}
\alias{get_Hhat}
\title{Adds 3-year average to harvest matrix}
\usage{
get_Hhat(dat)
}
\arguments{
\item{dat}{}
}
\value{
A matrix
}
\description{
This function creates a matrix containing harvest data for use in jags. If the SRA includes years without statewide harvest 
survey data a 3-year average is used for the missing years.
}
\examples{
Ha.hat <- get_Hhat(Ha)
Hd.hat <- get_hhat(Hd)

}
