% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_ids}
\alias{get_ids}
\title{Lookup tables for year, age and stock}
\usage{
get_ids(year_range = 1979:2017, age_id = c(age3 = 1, age4 = 2, age5 = 3,
  age678 = 4))
}
\arguments{
\item{year_range}{character vector with the indices of abundance included in the analysis.}

\item{age_id}{named vector of age names and age column position.}
}
\value{
writes 6 objects to R's Global Environment; year_id and stock_id where each object is a named vector where the names are jags array 
locations and the elements are informative. trib_id which is a named list with the tributaries comprising each stock. age_id where each object is a named vector where the names are informative
and the elements are jags array locations. Also, age_min and age_max the number of years between spawning and recruitment for the youngest and
oldest non-negligible age classes.
}
\description{
Creates lookup tables to switch between informative names and jags array locations.
}
\examples{
get_ids(1979:2017)

}
