% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_inits}
\alias{get_inits}
\title{Initial values for Jags model}
\usage{
get_inits()
}
\value{
A list of initial values
}
\description{
Function to generate random but constrainded initial values.
}
\examples{
get_inits()

}
