% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_profile}
\alias{get_profile}
\title{Creates a dataset for plotting OYP, ORP and EY plots}
\usage{
get_profile(post_dat, stock_name)
}
\arguments{
\item{post_dat}{An mcmc object with nodes lnalpha, beta, S.msy, lnalpha.c from which you want to simulate SR relationships.}

\item{stock_name}{A character element specifying the stock to plot.}
}
\value{
A data.frame
}
\description{
This function creates a dataframe that can be used by plot_profile()
}
\examples{
get_ids()
get_profile(post, stock_id[1])
get_profile(post, "East Susitna")
lapply(stock_id, get_profile, post_dat = post)

}
