% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_profileprobs}
\alias{get_profileprobs}
\title{OYP, ORP and OFP probabilities}
\usage{
get_profileprobs(profile_dat, goal_range = NULL, probs = NULL)
}
\arguments{
\item{profile_dat}{Output of the get_profile function}

\item{goal_range}{A vector with two element c(lower_bound, upper_bound). Defaults to NULL.}

\item{probs}{A list containing a 2 element vector. The vector is named after the profile of interest (e.g. "OYP90" of "ORP70") while 
the vector contains the desired probability at each bound c(lower_bound, upper_bound). Defaults to NULL.}
}
\value{
a dataframe with 2 rows.
}
\description{
Function can be used in two ways. Specifying a goal range will return the OYP, OFP and ORP probabilities of each end of the goal range. 
Specifying OYP or ORP probabilities will return the escapement sizes associated with those probabilities.
}
\examples{
get_profileprobs(profile, goal_range = c(10000, 20000))
get_profileprobs(profile, probs = list(OYP80 = c(.8, .3))

}
