% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_Rhat}
\alias{get_Rhat}
\title{Identify parameters with questionable convergence.}
\usage{
get_Rhat(post, cutoff = 1.1)
}
\arguments{
\item{post}{Jags posterior object}

\item{cutoff}{returns parameters with an Rhat that exceeds the cutoff.}
}
\value{
a list with 2 elements.  The parameters with a Rhat exceeding the cutoff and the Rhat values forming the 90%+ quartiles.
}
\description{
Return parameters from a RJags posterior object with an Rhat that exceed a user specified cutoff.
}
\examples{
get_Rhat(post)

}
