% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lt500}
\alias{lt500}
\title{Chinook salmon size data}
\format{data.frame
  \describe{
  \item{year}{Year collected}
  \item{age}{Age class}
  \item{n_small}{Number of fish sampled less than 500mm MEFT}
  \item{n}{Total Number of fish sampled}
  \item{p)small}{Percentage of fish less than 500mm MEFT}
  }}
\usage{
lt500
}
\description{
Number of Chinook salmon by size class (500mm MEFT boundary)
}
\keyword{datasets}
