% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mr}
\alias{mr}
\title{Mark-recapture abundance estimates}
\format{list
  \describe{
  \item{mr}{Abundance estimates}
  \item{cv_mr}{Coefficient of variation for abundance estimates}
  }}
\usage{
mr
}
\description{
Mark-recapture abundance estimates by stock
}
\keyword{datasets}
