% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_age}
\alias{plot_age}
\title{Plots of composition and abundance by age}
\usage{
plot_age(post_dat)
}
\arguments{
\item{post_dat}{The posterior object from the SRA model of class jagsUI}
}
\value{
A figure
}
\description{
Faceted plot of age at maturity, age composition and total run by age.  Observed age composition is also plotted.
}
\examples{
plot_age(post)

}
