% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_countprofile}
\alias{plot_countprofile}
\title{OYP, ORP and/or OFP plots based on aerial counts}
\usage{
plot_countprofile(profile_dat, limit = NULL, goal_range = NA,
  profiles = c("OYP", "ORP", "OFP"))
}
\arguments{
\item{profile_dat}{Output of the get_profile function}

\item{limit}{Upper bound of spawners for plot. Default (NULL) will use the maximum observed count.}

\item{goal_range}{A vector with two element c(lower_bound, upper_bound). Defaults to NA.}

\item{profiles}{the profiles to plot as a character vector e.g. c("OYP", "OFP").  Defaults to c("OYP", "ORP", "OFP").}
}
\value{
A figure
}
\description{
Produces a faceted plot of OYP, ORP or OFP with an overlay of the proposed goal range.
}
\examples{
plot_countprofile(get_countprofile(post, 200000), c("OYP", "OFP"))

}
