% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ey}
\alias{plot_ey}
\title{Expected sustained yield plot}
\usage{
plot_ey(profile_dat, limit = NULL, rug = TRUE, goal_range = NA)
}
\arguments{
\item{profile_dat}{Output of the profile data function}

\item{limit}{Upper bounds for plot c(xmax, ymax). Default (NULL) will pick bounds from the data.}

\item{rug}{Show scaled statewide goal ranges. Defaults to TRUE.}

\item{goal_range}{A vector with two element c(lower_bound, upper_bound). Defaults to NA.}
}
\value{
A figure
}
\description{
Expected sustained yield plot with 50 percent confidence ribbon
}
\examples{
get_ids()
plot_ey(get_profile(post, stock_id[1]), c(5000, 9000))
profiles <- lapply(stock_id, get_profile, post_dat = post)
lapply(profiles, plot_ey)

}
