% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_fit}
\alias{plot_fit}
\title{Model fit plots}
\usage{
plot_fit(post_dat, stock_name)
}
\arguments{
\item{post_dat}{The posterior object from the SRA model of class jagsUI}

\item{stock_name}{A character element specifying the stock to plot.}
}
\value{
A figure
}
\description{
Produces a faceted plot of Escapement and Inriver Run with the appropriately scaled indices of abundance that were used as inputs to the model.
}
\examples{
get_ids()
plot_fit(post, stock_id[1])
plot_fit(post, "East Susitna")
lapply(stock_id, plot_fit, post_dat = post)

}
