% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_horse}
\alias{plot_horse}
\title{Horsetail plot of plausible spawn-recruit relationships}
\usage{
plot_horse(post_dat, stock_name)
}
\arguments{
\item{post_dat}{SRA model mcmc.list output}

\item{stock_name}{A character element specifying the stock to plot.}
}
\value{
A figure
}
\description{
Produces a horsetail plot of the mean Spawn-Recruit relationship.  Plot also shows 40 plausible Spawn-Recruit relationships in the background and Spawner and Recruit estimates with associated 90% CIs.
}
\examples{
get_ids()
plot_horse(post, stock_id[1])
plot_horse(post, "East Susitna")
lapply(stock_id, plot_horse, post_dat = post)

}
