% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_rickeryear}
\alias{plot_rickeryear}
\title{Ricker Curves with annual productivity}
\usage{
plot_rickeryear(post_dat, stock_name)
}
\arguments{
\item{post_dat}{The posterior object from the SRA model of class jagsUI}

\item{stock_name}{A character element specifying the stock to plot.}
}
\value{
A figure
}
\description{
logRhat.y <- log(S.y) + lnalpha - beta * S.y 
logresid.y <- log(R.y) - logRhat.y
lnalpha.y <- lnalpha + logresid.y
}
\examples{
get_ids()
plot_rickeryear(post, stock_id[1])
plot_rickeryear(post, "East Susitna")
lapply(stock_id, plot_rickeryear, post_dat = post)

}
