% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_state}
\alias{plot_state}
\title{State Variable Plot}
\usage{
plot_state(post_dat, stock_name, rp = NULL)
}
\arguments{
\item{post_dat}{The posterior object from the SRA model of class jagsUI}

\item{stock_name}{A character element specifying the stock to plot.}

\item{rp}{NULL omits biological refence points, "msr" or "msy" include reference points for harvest rate and spawning abundance.}
}
\value{
A figure
}
\description{
Produces a faceted plot of escapement, recruitment, total run, Ricker residuals and harvest rate plotted with 95% confidence envelopes.
}
\examples{
plot_state(post, stock_id[1])
plot_state(post, "East Susitna")
lapply(stock_id, plot_state, post_dat = post)

}
