% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_statepairs}
\alias{plot_statepairs}
\title{Ricker Posterior correlations plot}
\usage{
plot_statepairs(post_dat, plot)
}
\arguments{
\item{plot}{"bystock" a separate plot of each stock, "byparam" for a separate plot for each Ricker parameter}

\item{dat_post}{posterior object}
}
\value{
A figure
}
\description{
Pairs plot for beta, ln(alpha), phi, S.msy and sigma.white
}
\examples{
plot_statepairs(post)

}
