% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_stock}
\alias{plot_stock}
\title{Plots of composition by stock}
\usage{
plot_stock(input_dat, post_dat)
}
\arguments{
\item{input_dat}{telemetry data matrix}

\item{post_dat}{The posterior object from the SRA model of class jagsUI}
}
\value{
A figure
}
\description{
Faceted plot of stock composition for Susitna Drainage. (Need to add observations)
}
\examples{
plot_stock(telemetry, post)

}
