% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_Swgoals}
\alias{plot_Swgoals}
\title{Plot of escapement vrs. proposed goals faceted by stock}
\usage{
plot_Swgoals(post_dat, goal_range)
}
\arguments{
\item{post_dat}{The posterior object from the SRA model of class jagsUI}

\item{goal_range}{A data frame with columns "stock", "lb" the lower bound of the goal range and "ub" the upper bound of the goal range.}
}
\value{
A figure
}
\description{
Produces a faceted plot of model estimated escapement with 95% CI error bars overlain by proposed goal ranges for each stock.
}
\examples{
plot_Swgoals(post, goals)

}
