% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_theta}
\alias{plot_theta}
\title{Plots of Aerial Survey detectability}
\usage{
plot_theta(post_dat)
}
\arguments{
\item{post_dat}{The posterior object from the SRA model of class jagsUI}
}
\value{
A figure
}
\description{
Estimated and observed aerial survey detectability in the Susitna River drainage.
}
\examples{
plot_theta(post)

}
