% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table_airerror}
\alias{table_airerror}
\title{Table of Aerial Survey Observability and Standard Deviations.}
\usage{
table_airerror(post_dat)
}
\arguments{
\item{post_dat}{The posterior object from the SRA model of class jagsUI}
}
\value{
A table
}
\description{
Produces a table of air survey observability and standard deviations for each flown tributary.
}
\examples{
table_airerror(post)

}
