% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table_brood}
\alias{table_brood}
\title{Brood table}
\usage{
table_brood(post_dat, stock)
}
\arguments{
\item{post_dat}{The posterior object from the SRA model of class jagsUI}

\item{stock}{numeric. See stock_id for numeric codes.}
}
\value{
A tibble
}
\description{
Produces an brood tables based on the posterior means form the state space model.  Most of the table functions in this package are intended
 for output into word via markdown.  This one produces a tibble object that can be output to excel using WriteXLS::WriteXLS.
}
\examples{
table_brood(post_dat, stock)

}
