% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table_params}
\alias{table_params}
\title{Table of SR analysis paramerater estimates}
\usage{
table_params(post_dat, stocks = 1:4)
}
\arguments{
\item{post_dat}{The posterior object from the SRA model of class jagsUI}
}
\value{
A table
}
\description{
Produces a table of paramerater estimates for the SRA.
}
\examples{
table_params(post)

}
