% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table_state}
\alias{table_state}
\title{State Variable Table}
\usage{
table_state(post_dat, display)
}
\arguments{
\item{post_dat}{The posterior object from the SRA model of class jagsUI}

\item{display}{"bystock" for a list of state parameter estimates for each stock and "bystate" for 
a list of stock parameter estimates for each state.}
}
\value{
A table
}
\description{
Produces a table of escapement, recruitment, total run, and inriver run along with cv's.
}
\examples{
table_state(post, "bystate")

}
