% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table_stock}
\alias{table_stock}
\title{Stock composition table}
\usage{
table_stock(post_dat)
}
\arguments{
\item{post_dat}{The posterior object from the SRA model of class jagsUI}
}
\value{
A table
}
\description{
Produces a table of stock composition by drainage along with cv's.
}
\examples{
table_stock(post)

}
