% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{telemetry}
\alias{telemetry}
\title{Telemetry data}
\format{list
 \describe{
  \item{East Susitna}{A matrix of radio telemetry final destinations in the East Susitna area}
  \item{N_East Susitna}{Row sums of the East Susitna matrix}
  \item{Talkeetna}{A matrix of radio telemetry final destinations in the Talkeetna drainage}
  \item{N_Talkeetna}{Row sums of the Talkeetna matrix}
  \item{Yentna}{A matrix of radio telemetry final destinations in the Yentna drainage}
  \item{N_Yentna}{Row sums of the Yentna matrix}
  }}
\usage{
telemetry
}
\description{
Radio Telemetry final destinations by Stock
}
\keyword{datasets}
