% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hm}
\alias{Hm}
\title{Marine Harvest data}
\format{tibble}
\usage{
Hm
}
\description{
Commercial and Substance harvest of Chinook salmon in North Cook inlet. Apportioned by GSI.
}
\keyword{datasets}
