% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_state}
\alias{plot_state}
\title{State Variable Plot}
\usage{
plot_state(post_dat, display, rp = NULL)
}
\arguments{
\item{post_dat}{The posterior object from the SRA model of class jagsUI}

\item{display}{A character element. One of the stock group names; "Deshka", "East Susitna", "Talkeetna", "Yentna", 
or state variable names: "Escapement", "Total Run", "Recruitment", "Harvest Rate", "Ricker Residuals". 
Choosing a stock group name will facet 5 state variables. Choosing a state variable will facet 4 stock groups.}

\item{rp}{NULL omits biological refence points, "msr" or "msy" include reference points for harvest rate and spawning abundance.}
}
\value{
A figure
}
\description{
Produces a plot of escapement, recruitment, total run, Ricker residuals and/or harvest rate plotted with 95% confidence envelopes
 for one or all of the stock groups.
}
\examples{
plot_state(post, stock_id[1])
plot_state(post, "East Susitna")
lapply(stock_id, plot_state, post_dat = post)

}
