% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_stock}
\alias{plot_stock}
\title{Plots of composition by stock}
\usage{
plot_stock(input_dat, post_dat, plot_stocks = c("East Susitna",
  "Talkeetna", "Yentna"))
}
\arguments{
\item{input_dat}{telemetry data matrix}

\item{post_dat}{The posterior object from the SRA model of class jagsUI}

\item{plot_stocks}{The stocks to include in the plot. The default c("East Susitna", "Talkeetna", "Yentna").}
}
\value{
A figure
}
\description{
Faceted plot of stock composition for Susitna Drainage.
}
\examples{
plot_stock(telemetry, post)

}
